using $lib in imports and moving stuff in lib
