+++ src/lib/components/FallingEmojis.svelte
