+++ src/lib/components/Favicons.svelte
