+++ src/lib/components/Nav.svelte
