+++ src/lib/speech.js
