+++ src/routes/404.html.svelte
@@ -1,6 +1,6 @@
 <script>
-  import TitleBar from '../components/TitleBar.svelte';
-  import Favicons from '../components/Favicons.svelte';
+  import TitleBar from '$lib/components/TitleBar.svelte';
+  import Favicons from '$lib/components/Favicons.svelte';
 </script>
 
 <svelte:head>
