+++ src/routes/future/[slug].svelte
@@ -25,7 +25,7 @@
 </script>
 
 <script>
-  import Post from '../../components/Post.svelte';
+  import Post from '$lib/components/Post.svelte';
 
   export let post;
 </script>
