+++ src/routes/rss.js
@@ -1,5 +1,5 @@
-import { getPosts } from './_posts.js';
-import { siteUrl } from '../stores/_config.js';
+import { getPosts } from '$lib/posts.js';
+import { siteUrl } from '$lib/stores/config.js';
 
 const renderXmlRssFeed = (posts) => `<?xml version="1.0" encoding="UTF-8" ?>
 <rss xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:content="http://purl.org/rss/1.0/modules/content/" xmlns:atom="http://www.w3.org/2005/Atom" version="2.0">
